chore: moved utils to separate files
